requirements.txt
asgiref==3.9.1
Django==4.2
django-cors-headers==4.0.0
djangorestframework==3.15.0
djangorestframework_simplejwt==5.5.1
gunicorn==20.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
setuptools==80.9.0
sqlparse==0.5.3
